提交结构体练习第二遍

+++ empty/23.go
@@ -6,7 +6,7 @@ func main() {
 
 type Cat struct {
 	Name  string
-	Age   float64
+	Age   int
 	Color string
 }
 
@@ -15,4 +15,4 @@ type Cat struct {
 // 2.使用自定义 json 序列号的 key ，将实例化结构体转为 json
 
 // 3.将字符串 json 转为实例化结构体
-// 	str1 := `{"Name":"小黄","Age":1,"Color:"yellow""}`
+// 	str1 := `{"Name":"小黄","Age":1,"Color":"yellow"}`
